сменил имя главного каталога проекта и актуализировал список библиотек в requirements.txt
