+++ README.md
@@ -1,6 +1,10 @@
 # Backend
 Репозиторий бакэнда
 
+python manage.py makemigrations:
+python manage.py migrate:
+python manage.py createsuperuser: admin admin
+python manage.py collectstatic:
 python manage.py add_themes:
     команда для загрузки стандартного набора тематик.
     запустить при первоначальной сборке проекта.
