--- requarement.txt
@@ -1,3 +1,0 @@
-Django==5.0.3
-djangorestframework==3.15.1
-drf-spectacular==0.27.0
